requirements.txt
python-dotenv
discord.py
discord
peewee
google-genai
rich
translate
openai
requests
simpy
matplotlib
beautifulsoup4
sympy
aiofiles
selenium
numpy
selenium_stealth
fake_useragent
pydub
pydantic
undetected-chromedriver